+++ bashlib64.bash
@@ -1622,7 +1622,7 @@ function bl64_check_module_setup() {
 #######################################
 # BashLib64 / Module / Setup / Interact with container engines
 #
-# Version: 1.3.0
+# Version: 1.4.0
 #######################################
 
 #######################################
@@ -1642,8 +1642,16 @@ function bl64_check_module_setup() {
 function bl64_cnt_setup() {
   bl64_dbg_lib_show_function
 
-  bl64_cnt_set_command &&
+  bl64_cnt_set_command || return $?
+
+  if [[ ! -x "$BL64_CNT_CMD_DOCKER" && ! -x "$BL64_CNT_CMD_PODMAN" ]]; then
+    bl64_msg_show_error "$_BL64_CNT_TXT_NO_CLI (docker or podman)"
+    # shellcheck disable=SC2086
+    return $BL64_LIB_ERROR_APP_MISSING
+  else
     BL64_CNT_MODULE="$BL64_LIB_VAR_ON"
+  fi
+
 }
 
 #######################################
@@ -1682,7 +1690,7 @@ function bl64_cnt_set_command() {
 #######################################
 # BashLib64 / Module / Functions / Interact with container engines
 #
-# Version: 1.4.0
+# Version: 1.5.0
 #######################################
 
 #######################################
@@ -1714,10 +1722,6 @@ function bl64_cnt_login_file() {
     bl64_cnt_docker_login "$user" "$BL64_LIB_DEFAULT" "$file" "$registry"
   elif [[ -x "$BL64_CNT_CMD_PODMAN" ]]; then
     bl64_cnt_podman_login "$user" "$BL64_LIB_DEFAULT" "$file" "$registry"
-  else
-    bl64_msg_show_error "$_BL64_CNT_TXT_NO_CLI (docker or podman)"
-    # shellcheck disable=SC2086
-    return $BL64_LIB_ERROR_APP_MISSING
   fi
 }
 
@@ -1749,10 +1753,6 @@ function bl64_cnt_login() {
     bl64_cnt_docker_login "$user" "$password" "$BL64_LIB_DEFAULT" "$registry"
   elif [[ -x "$BL64_CNT_CMD_PODMAN" ]]; then
     bl64_cnt_podman_login "$user" "$password" "$BL64_LIB_DEFAULT" "$registry"
-  else
-    bl64_msg_show_error "$_BL64_CNT_TXT_NO_CLI (docker or podman)"
-    # shellcheck disable=SC2086
-    return $BL64_LIB_ERROR_APP_MISSING
   fi
 }
 
@@ -1859,10 +1859,6 @@ function bl64_cnt_run_interactive() {
     bl64_cnt_docker_run_interactive "$@"
   elif [[ -x "$BL64_CNT_CMD_PODMAN" ]]; then
     bl64_cnt_podman_run_interactive "$@"
-  else
-    bl64_msg_show_error "$_BL64_CNT_TXT_NO_CLI (docker or podman)"
-    # shellcheck disable=SC2086
-    return $BL64_LIB_ERROR_APP_MISSING
   fi
 }
 
@@ -2014,10 +2010,6 @@ function bl64_cnt_build() {
     bl64_cnt_docker_build "$file" "$tag"
   elif [[ -x "$BL64_CNT_CMD_PODMAN" ]]; then
     bl64_cnt_podman_build "$file" "$tag"
-  else
-    bl64_msg_show_error "$_BL64_CNT_TXT_NO_CLI (docker or podman)"
-    # shellcheck disable=SC2086
-    return $BL64_LIB_ERROR_APP_MISSING
   fi
 }
 
@@ -2104,10 +2096,6 @@ function bl64_cnt_push() {
     bl64_cnt_docker_push "$source" "$destination"
   elif [[ -x "$BL64_CNT_CMD_PODMAN" ]]; then
     bl64_cnt_podman_push "$source" "$destination"
-  else
-    bl64_msg_show_error "$_BL64_CNT_TXT_NO_CLI (docker or podman)"
-    # shellcheck disable=SC2086
-    return $BL64_LIB_ERROR_APP_MISSING
   fi
 }
 
@@ -2185,10 +2173,6 @@ function bl64_cnt_pull() {
     bl64_cnt_docker_pull "$source"
   elif [[ -x "$BL64_CNT_CMD_PODMAN" ]]; then
     bl64_cnt_podman_pull "$source"
-  else
-    bl64_msg_show_error "$_BL64_CNT_TXT_NO_CLI (docker or podman)"
-    # shellcheck disable=SC2086
-    return $BL64_LIB_ERROR_APP_MISSING
   fi
 }
 
@@ -2271,10 +2255,6 @@ function bl64_cnt_tag() {
     bl64_cnt_docker_tag "$source" "$target"
   elif [[ -x "$BL64_CNT_CMD_PODMAN" ]]; then
     bl64_cnt_podman_tag "$source" "$target"
-  else
-    bl64_msg_show_error "$_BL64_CNT_TXT_NO_CLI (docker or podman)"
-    # shellcheck disable=SC2086
-    return $BL64_LIB_ERROR_APP_MISSING
   fi
 }
 
@@ -2324,6 +2304,74 @@ function bl64_cnt_podman_tag() {
     tag \
     "$source" \
     "$target"
+}
+
+#######################################
+# Runs a container image
+#
+# Arguments:
+#   $@: arguments are passes as-is
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+function bl64_cnt_run() {
+  bl64_dbg_lib_show_function "$@"
+
+  if [[ -x "$BL64_CNT_CMD_DOCKER" ]]; then
+    bl64_cnt_docker_run "$@"
+  elif [[ -x "$BL64_CNT_CMD_PODMAN" ]]; then
+    bl64_cnt_podman_run "$@"
+  fi
+}
+
+#######################################
+# Command wrapper: podman run
+#
+# * Provides verbose and debug support
+#
+# Arguments:
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+
+function bl64_cnt_podman_run() {
+  bl64_dbg_lib_show_function "$@"
+
+  bl64_cnt_run_podman \
+    run \
+    --rm \
+    "$@"
+}
+
+#######################################
+# Command wrapper: docker run
+#
+# * Provides verbose and debug support
+#
+# Arguments:
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+
+function bl64_cnt_docker_run() {
+  bl64_dbg_lib_show_function "$@"
+
+  bl64_cnt_run_docker \
+    run \
+    --rm \
+    "$@"
+
 }
 
 #######################################